Exception treatment delete

+++ src/main/java/com/projectSpringJPA/project/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.projectSpringJPA.project.entities.User;
 import com.projectSpringJPA.project.repositories.UserRepository;
+import com.projectSpringJPA.project.services.exceptions.DatabaseException;
 import com.projectSpringJPA.project.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,18 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		ur.deleteById(id);
+	    try {
+	        if (ur.existsById(id)) {
+	            ur.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    }catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }
+	    catch (DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	public User update(Long id, User user) {
